Move everything related to verification to internal/verification package
