+++ internal/verification/verifier.go
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package verifier provides a function for verifying a SLSA provenance file.
-package verifier
+// Package verification provides a function for verifying a SLSA provenance file.
+package verification
 
 import (
 	"fmt"
@@ -23,18 +23,11 @@ import (
 	"go.uber.org/multierr"
 )
 
-// ProvenanceVerifier defines an interface with a single method `Verify` for
-// verifying provenances.
-type ProvenanceVerifier interface {
-	// Verifies a provenance.
-	Verify() error
-}
-
 // ProvenanceIRVerifier verifies a provenance against a given reference, by verifying
 // all non-empty fields in got using fields in the reference values. Empty fields will not be verified.
 type ProvenanceIRVerifier struct {
 	Got  *common.ProvenanceIR
-	Want *common.ReferenceValues
+	Want *ReferenceValues
 }
 
 // Verify verifies an instance of ProvenanceIRVerifier by comparing its Got and Want fields.
